Remove cli and two-step abstraction
