+++ config-sample.json
@@ -1,0 +1,15 @@
+{
+  "servers": {
+    "everything": {
+      "type": "stdio",
+      "command": "npx",
+      "args": [
+        "-y",
+        "@modelcontextprotocol/server-everything"
+      ],
+      "env": {
+        "DEBUG": "true"
+      }
+    }
+  }
+}
