--- configuration/servers.yaml
@@ -1,7 +1,0 @@
-servers:
-  sequentialthinking:
-    type: stdio
-    command: npx
-    args:
-      - "-y"
-      - "@modelcontextprotocol/server-sequential-thinking"
