+++ connector-definition/.hasura-connector/connector-metadata.yaml
@@ -3,18 +3,6 @@ ndcSpecGeneration: v0.2
 packagingDefinition:
   type: ManagedDockerBuild
 supportedEnvironmentVariables: []
-commands:
-  update:
-    type: Dockerized
-    dockerImage: ghcr.io/hasura/mcp-connector:v0.1.0
-    commandArgs:
-      - "mcp-connector-cli"
-      - "update"
-      - "--outfile"
-      - "/etc/connector/configuration.json"
-cliPlugin:
-  type: Docker
-  dockerImage: ghcr.io/hasura/mcp-connector:v0.1.0
 dockerComposeWatch:
   # Rebuild the container if a new package restore is required because Cargo.toml or Cargo.lock changed
   - path: Cargo.toml
