+++ justfile
@@ -16,10 +16,6 @@ build-release:
 serve:
     cargo run --bin mcp-connector -- serve --configuration configuration
 
-# Generate configuration file
-generate-config:
-    cargo run --bin mcp-connector-cli -- --configuration configuration update --outfile configuration/configuration.json
-
 # Format the code
 format:
     cargo fmt
