--- servers-sample.yaml
@@ -1,9 +1,0 @@
-servers:
-  everything:
-    type: stdio
-    command: npx
-    args:
-      - "-y"
-      - "@modelcontextprotocol/server-everything"
-    env:
-      DEBUG: "true"
